Refactor: Move ArrayLookupTable tests to dedicated file
